Renamed folder to 'lib'
