+++ pages/dev/map/layers/lib/raster-map.ts
