+++ pages/dev/map/layers/lib/utils.ts
